# public/swager/crm/api-docs.yaml
openapi: 3.0.0
info:
  title: 'Суши Фокс REST API для CRM'
  description: ''
  contact:
    email: dev@sushifox.ru
  version: 0.0.1
servers:
  - url: 'http://api.sushifox.ru/crm'
    description: 'API sushi fox server'
  - url: 'http://localhost/crm'
    description: 'local server'
  - url: 'http://sushifox.local/crm'
    description: 'local server 2'
components:
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes
security:
  - bearerAuth: [ ]
paths:
...
  /transactions/{id}/wait:
    patch:
      tags:
        - 'Транзакции'
      summary: 'Перевести транзакцию в режим ожидания'
      description: 'Перевести транзакцию в режим ожидания'
      operationId: waitTransaction
      parameters:
        - name: id
          in: path
          description: 'Внутрений индификатор транзакции'
          required: true
          schema:
            type: string
      responses:
        200:
          description: 'Успешный ответ'
          content:
            application/json:
              examples:
                - $ref: ../api-example/curd/edit.json
        422:
          description: 'Неправильные данные'
          content:
            application/json:
              examples:
                - $ref: ../api-example/invalid_request.json
        401:
          description: Неавторизованный
          content:
            application/json:
              examples:
                - $ref: ../api-example/auth_fail.json
        403:
          description: 'Не разрешено! В доступе отказано'
          content:
            application/json:
              examples:
                - $ref: ../api-example/access_denied.json
...